Use async (#4)

+++ src/dto/filter_types.rs
@@ -3,12 +3,12 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct EpochTimestampRange {
-    max: u128,
-    min: u128,
+    max: i64,
+    min: i64,
 }
 
 impl EpochTimestampRange {
-    pub fn new(min: u128, max: u128) -> EpochTimestampRange {
+    pub fn new(min: i64, max: i64) -> EpochTimestampRange {
         EpochTimestampRange { min, max }
     }
 }
@@ -16,12 +16,12 @@ impl EpochTimestampRange {
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct IntegerRange {
-    max: u64,
-    min: u64,
+    max: i64,
+    min: i64,
 }
 
 impl IntegerRange {
-    pub fn new(min: u64, max: u64) -> IntegerRange {
+    pub fn new(min: i64, max: i64) -> IntegerRange {
         IntegerRange { min, max }
     }
 }
